Removed Handler.php and introduced ExceptionMapper.php to handle mapping of exceptions to the right responses
